+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2.1.10 (Unreleased)
 - [Enhancement] Introduce `connection.client.rebalance_callback` event for instrumentation of rebalances.
 - [Fix] Always try to unsubscribe short-lived consumers used throughout the system, especially in the admin APIs.
+- [Fix] Always unsubscribe long-lived consumers prior to shutdown.
+- [Fix] Add missing `connection.client.poll.error` error type reference.
 
 ## 2.1.9 (2023-08-06)
 - **[Feature]** Introduce ability to customize pause strategy on a per topic basis (Pro).